Add : Forgot to add Weeks variable to right-click on previous commit.
Fix :
When saving a webpage (Ctrl+S) without the page name in the URL, filtering on the .htm extension is now working.
When saving a page with its name present on the URL, it still filters on the real page name without the ".htm" which Firefox add on the file explorer.
Example 1:
http://domain.com/ -> Firefox will create a filename based on WebpageTitle.htm  and the filter will check webpageTitle.htm
Example 2:
http://domain.com/index.php -> Same page, but Firefox will create a filename based on the filename like index.php.htm, but the filters will check only index.php

To filter all webpages, create a regular expression filter like /.*\.(html?|php)$/